refactor: extract context menu page functionality to hook
