+++ app/client/src/ee/pages/AppIDE/hooks/useGenPageItems.tsx
@@ -1,0 +1,1 @@
+export * from "ce/pages/AppIDE/hooks/useGenPageItems";
